improve(web): use global variable
