+++ web/src/components/MsgPublish.vue
@@ -157,7 +157,6 @@ import ClickOutside from 'vue-click-outside'
 import jump from 'jump.js'
 import Editor from '@/components/Editor.vue'
 import KeyValueEditor from '@/components/KeyValueEditor.vue'
-import { MessageModel } from '../views/connections/types'
 import convertPayload from '@/utils/convertPayload'
 import { getMessageId } from '@/utils/idGenerator'
 import _ from 'lodash'
