+++ web/src/components/SubscriptionsList.vue
@@ -110,7 +110,6 @@ import { updateConnection } from '@/utils/api/connection'
 import { defineColors, getRandomColor } from '@/utils/colors'
 import LeftPanel from '@/components/LeftPanel.vue'
 import MyDialog from '@/components/MyDialog.vue'
-import { ConnectionModel } from '../views/connections/types'
 
 @Component({
   components: {
