+++ web/src/utils/getFiles.ts
@@ -1,5 +1,4 @@
 import { Notification } from 'element-ui'
-import { SSLPath, SSLContent } from '@/views/connections/types'
 
 export const getSSLFile = (sslPath: SSLPath): SSLContent | undefined => {
   const { ca, cert, key } = sslPath
